requirements.txt
#mysql
mysql-connector-python==8.0.32
streamlit==1.21.0
streamlit-chat==0.0.2.2
sqlvalidator==0.0.20
streamlit_option_menu==0.3.2
pyyaml==6.0
openai